web: disable background feature flags evaluation
